+++ src/main/java/com/recipe/app/src/recipe/infra/blog/BlogRecipeEntity.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.recipe.infra;
+package com.recipe.app.src.recipe.infra.blog;
 
 import com.recipe.app.common.entity.BaseEntity;
 import com.recipe.app.src.recipe.domain.BlogRecipe;
