+++ src/main/java/com/recipe/app/src/recipe/infra/youtube/YoutubeRecipeJpaRepository.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.recipe.infra;
+package com.recipe.app.src.recipe.infra.youtube;
 
 import org.springframework.data.repository.CrudRepository;
